Delete index test and migrate to component-specific unit tests
