+++ src/app/containers/EpisodeList/RecentVideoEpisodes/index.test.jsx
@@ -20,7 +20,7 @@ const RecentVideoEpisodesWithContext = ({ episodes, isAmp = false }) => (
   </ServiceContextProvider>
 );
 
-describe('RecentAudioEpisodes', () => {
+describe('Recent Video Episodes', () => {
   shouldMatchSnapshot(
     'should render video episodes correctly',
     <RecentVideoEpisodesWithContext episodes={afrique} />,
@@ -41,6 +41,30 @@ describe('RecentAudioEpisodes', () => {
     );
 
     expect(container.querySelectorAll('li').length).toEqual(3);
+  });
+  it('should not render a list when there is only one episode', () => {
+    const { queryByRole } = render(
+      <RecentVideoEpisodesWithContext episodes={[afrique[0]]} />,
+    );
+
+    expect(queryByRole('list')).not.toBeInTheDocument();
+    expect(queryByRole('listitem')).not.toBeInTheDocument();
+  });
+
+  it('should render the brand title', () => {
+    const { getAllByText } = render(
+      <RecentVideoEpisodesWithContext episodes={afrique} />,
+    );
+
+    expect(getAllByText('BBC Info')[0]).toBeInTheDocument();
+  });
+
+  it('should render the episode title', () => {
+    const { getByText } = render(
+      <RecentVideoEpisodesWithContext episodes={afrique} />,
+    );
+
+    expect(getByText('Oui, je suis le chef')).toBeInTheDocument();
   });
 
   it('should render the list item links', async () => {
@@ -79,6 +103,47 @@ describe('RecentAudioEpisodes', () => {
 
     expect(hiddenDuration).toBeDefined();
     expect(hiddenDuration).toContainHTML('15:00');
+  });
+
+  it('should correctly handle images', () => {
+    const { container } = render(
+      <RecentVideoEpisodesWithContext episodes={afrique} />,
+    );
+
+    expect(
+      container.querySelector(`img[src='${afrique[0].image}']`),
+    ).toBeInTheDocument();
+    expect(
+      container.querySelector(`img[alt='${afrique[0].altText}']`),
+    ).toBeInTheDocument();
+  });
+
+  it('should render the media indicator', () => {
+    const { container } = render(
+      <RecentVideoEpisodesWithContext episodes={afrique} />,
+    );
+    const svgs = container.querySelectorAll('svg');
+
+    expect(svgs).toHaveLength(3);
+  });
+
+  it('should include the data-e2e attribute if passed', () => {
+    const { container } = render(
+      <RecentVideoEpisodesWithContext
+        episodes={afrique}
+        ulProps={{ 'data-e2e': 'recent-episode-list' }}
+        liProps={{ 'data-e2e': 'recent-episode-list-item' }}
+      />,
+    );
+
+    expect(container.querySelector('ul')).toHaveAttribute(
+      'data-e2e',
+      'recent-episodes-list',
+    );
+    expect(container.querySelector('li')).toHaveAttribute(
+      'data-e2e',
+      'recent-episodes-list-item',
+    );
   });
 
   describe('on amp', () => {
